api/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
cloudinary
python-multipart
pydantic-settings
email-validator
Jinja2
